refector: handleRaceTimeInput 함수에서 게임 시작 부분 분리
