+++ js/controller/handleRaceTimeInput.js
@@ -1,32 +1,15 @@
 import { store } from "../model.js";
-import { carNameRender } from "./carNameRender.js";
-import { renderWinner } from "./renderWinner.js";
-import {
-  randomGoStop,
-  paintSpinnerIcon,
-  deleteSpinnerIcon,
-} from "./gameProcess.js";
+import { startGame } from "./startGame.js";
 
 export const handleRaceTimeInput = (event) => {
   event.preventDefault();
   const raceTimeInput = document.querySelector("#raceTimeInput");
   const raceTimeSubmitButton = document.querySelector("#raceTimeSubmitButton");
 
-  store.raceTime = Number(raceTimeInput.value);
+  store.raceTime = parseInt(raceTimeInput.value, 10);
 
   raceTimeInput.disabled = true;
   raceTimeSubmitButton.disabled = true;
 
-  carNameRender();
-  paintSpinnerIcon();
-
-  const raceInterval = setInterval(() => {
-    randomGoStop();
-    store.raceTime--;
-    if (store.raceTime === 0) {
-      clearInterval(raceInterval);
-      deleteSpinnerIcon();
-      renderWinner();
-    }
-  }, 2000);
+  startGame();
 };
